requirements.txt
aiohttp==3.8.4
aiohttp-socks==0.8.0
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.2
multidict==6.0.4
orjson==3.8.5
pydantic==1.10.4
python-dotenv==0.21.1
python-multipart==0.0.5
python-socks==2.2.0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
win32-setctime==1.1.0
yarl==1.8.2
